django_project/requirements.txt
django
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
slackclient
PyGithub
dnspython
pymongo
requests
schedule
